+++ init.sql
@@ -8,7 +8,7 @@ CREATE TABLE IF NOT EXISTS users (
 CREATE TABLE IF NOT EXISTS banks (
   id_bank BIGINT NOT NULL AUTO_INCREMENT PRIMARY KEY,
   name VARCHAR(255) NOT NULL,
-  initial_balance double,
+  initial_balance DECIMAL(12,2),
   user_id BIGINT,
   FOREIGN KEY (user_id) REFERENCES users (id_user)
 );
@@ -17,7 +17,7 @@ CREATE TABLE IF NOT EXISTS transactions (
   id_transaction BIGINT NOT NULL AUTO_INCREMENT PRIMARY KEY,
   transaction_type VARCHAR(255),
   date DATE,
-  amount double,
+  amount DECIMAL(12,2),
   description VARCHAR(255),
   category VARCHAR(255),
   origin_bank BIGINT,
